backend/requirements/common.txt
# Django
Django==3.2
django-environ==0.4.5  # manage environ
Pillow==8.2.0  # Imagefield

# REST API
djangorestframework==3.12.4  # REST
drf-spectacular  # api document
drf-access-policy  # permission
drf-nested-routers  # nested router
django-cachalot   # table cache
python-memcached   # memcached
django-filter==2.4.0

# WSGI
gunicorn==20.1.0